requirements.txt
# Core dependencies
langchain==0.1.14
openai
faiss-cpu
tiktoken

# PDF parsing
PyMuPDF==1.23.18  # fitz


# UI
streamlit==1.32.2

# Env handling
python-dotenv==1.0.1

# Utilities
numpy==1.26.4
pandas==2.2.1
